requirements.txt
# Core LangChain packages
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6
langchain-pinecone==0.0.1
langchain-huggingface==0.0.1
langchain-text-splitters==0.0.1

# Vector databases
faiss-cpu==1.7.4
pinecone-client==3.0.0
qdrant-client==1.7.0

# Document processing
pypdf==3.17.4
pymupdf==1.23.8

# ML/AI frameworks
torch==2.1.2
sentence-transformers==2.2.2
transformers==4.36.2

# LlamaIndex
# llama-index==0.9.48
# llama-index-core==0.9.48
# llama-index-embeddings-huggingface==0.1.4
# llama-index-llms-openai==0.1.6

# Utilities
python-dotenv==1.0.0

# HuggingFace Hub
# huggingface-hub==0.20.2
